requirements-offline.txt
mcp>=1.26.0
ansible-runner>=2.4.0
PyYAML>=6.0.2
